Introduce cc_ebpf rule to compile and embed ebpf code. (#18)
* Add bazel support to embed files into binaries directly.

This is step 2/N to have better toolchains, and embed eBPF
code into binaries.

The .bzl rule imported allows to link files directly into
binaries.

Example:
- You have a "trie_match.o" containing ebpf code.
- You use cc_embed(name = "trie_match", data = [":trie_match.o"])
- From your .cc file, you can now include "trie_match.h",
  and use std::string_view trie_match_o to access the raw bytes
  contained in the file.

This should work with separate targets as well, where :trie_match
is on its own generated by other rules.

* Add bazel support to embed files into binaries directly.

This is step 2/N to have better toolchains, and embed eBPF
code into binaries.

The .bzl rule imported allows to link files directly into
binaries.

Example:
- You have a "trie_match.o" containing ebpf code.
- You use cc_embed(name = "trie_match", data = [":trie_match.o"])
- From your .cc file, you can now include "trie_match.h",
  and use std::string_view trie_match_o to access the raw bytes
  contained in the file.

This should work with separate targets as well, where :trie_match
is on its own generated by other rules.

* Introduce cc_ebpf rule to compile and embed ebpf code.

Before this change:

eBPF code could be compiled with a normal cc_library target with
eBPF specific options. This, however, failed during the linking
stage, as it could not create an AR file.
Further, using the .o file was not really possible.

In this change:

- introduce a cc_build_ebpf rule, to compile ebpf code with the
  correct compile options. Unlike a cc_library rule, this stops
  right after compilation - does not attempt to create a library
  which would fail.

- introduce a cc_ebpf macro, that creates both a cc_build_ebpf
  rule, and a cc_embed rule, allowing to easily use ebpf code
  from C++ binaries.

- have a test to verify the rule works as expected.

Note that the test is still incomplete: there is no evidence that
the eBPF code is actually compiled correctly.

The vision here is that:
1) We'll add more tests as we add code to actually load and verify
   the eBPF code.
2) eBPF has a lot of sharp edges. I exepct the cc_ebpf rule will
   be updated over time to include more compiler flags, and automated
   checks [1].

[1]: see for example https://github.com/cilium/cilium/blob/master/Documentation/bpf.rst,
limits in instructions, constructs, and so on.

* Addressed Marcus comments.

* Introduce cc_ebpf rule to compile and embed ebpf code.

Before this change:

eBPF code could be compiled with a normal cc_library target with
eBPF specific options. This, however, failed during the linking
stage, as it could not create an AR file.
Further, using the .o file was not really possible.

In this change:

- introduce a cc_build_ebpf rule, to compile ebpf code with the
  correct compile options. Unlike a cc_library rule, this stops
  right after compilation - does not attempt to create a library
  which would fail.

- introduce a cc_ebpf macro, that creates both a cc_build_ebpf
  rule, and a cc_embed rule, allowing to easily use ebpf code
  from C++ binaries.

- have a test to verify the rule works as expected.

Note that the test is still incomplete: there is no evidence that
the eBPF code is actually compiled correctly.

The vision here is that:
1) We'll add more tests as we add code to actually load and verify
   the eBPF code.
2) eBPF has a lot of sharp edges. I exepct the cc_ebpf rule will
   be updated over time to include more compiler flags, and automated
   checks [1].

[1]: see for example https://github.com/cilium/cilium/blob/master/Documentation/bpf.rst,
limits in instructions, constructs, and so on.

* Added comment.

* Added more comments.