Implement error::Code and error::Status types
These classes encapsulate common error handling patterns in
exception-free C++. This implementation assumes that errors are
uncommon. In cases where errors are common, other types should be
used.

Note that this implementation differs from the standard Google style
util::Status in that it favors free functions over member functions.
This allows the same API (IsOk(), IsError()) to be used for querying
error::Code and error::Status.

A few basic macros are added to allow documentation and automatic
checking of invariants. When this code should be compiled and run is
left for future discussion.